pyproject.toml
[project]
name = "nexus-ai"
version = "0.1.0"
description = "Ultra-minimal multi-provider AI framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AI Provider SDKs - use ~= for compatible releases (allow minor updates)
    "openai~=1.12",
    "anthropic~=0.18",

    # CLI Framework - stable APIs, use >= with upper bounds
    "click>=8.1,<9.0",
    "rich>=13.0,<14.0",
    "prompt-toolkit>=3.0,<4.0",

    # Configuration & Utilities - stable APIs
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.0,<3.0",

    # HTTP & Retry - stable utilities
    "requests>=2.31,<3.0",
    "tenacity>=8.2,<9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.11",
    "hypothesis>=6.0",
    "responses>=0.24",
    "ruff>=0.1.0",
]

[project.scripts]
nexus = "nexus.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nexus"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
